Move `ErrorCode` to its own file and add more functionality, plus add the `ReturnType` struct.
These types will be used by `Syscalls` to communicate return values and errors back to drivers and client code.

Something I could use help on: naming. Let's use ErrorCode as the example, but both types struggle from this. There is ErrorCode, the struct that wraps the raw `usize` values. I also have a list of known error codes, but I didn't want to put them into an enum because we wouldn't use the enum type itself. Instead I tried to put them in a namespace, but the namespace name `error_code` clashes with the namespace they are contained in. I ended up renaming the namespace that `ErrorCode` is contained in to `struct_error_code` to avoid that collision, but that feels nasty.

Any ideas?